Further work on checkout functionality (payment views, def checkout and checkout.html)
working on def checkout function in payment views; 1. authenticated user, if authenticated user has shipping address already associated with his account, then it will be returned in a contect dictionary and during the next checkout it will be autofilled, if not the user needs to fill in the shipping form; 2. guest users need to fill in the shipping adress form; inserting values based on shipping address model and passing the context dictionary into checkout.html, so it can be rendered if the user is authenticated and we already have the shipping address in database