Further expanding def complete_checkout function in payment views py handling the checkout for users and guest users
Adding second part to def complete_checkout function different for 1. authenticated users with and without shipping data already saved and for 2. guest users;1. creation of a new order object in the database, associated with a logged in user, afterwards the new order item pk is assigned to order_id variable, the order_id links the order item to the specific order, serving as a foreignkey, the other foreign key connecting it to a logged in user; 2: for guest  user the process is almost the same with the only difference being the lack of a foreign key linking it to a user, since guest user is not a logged in and/or authenticated per its very definition; once the order and order item objects have been created the order_success variable is set to true and a dictionary is returned via JSON to Ajax script in checkout.html